pod_manager/requirements.txt
fastapi==0.110.0
kubernetes==29.0.0
redis==5.0.1
httpx==0.27.0
uvicorn[standard]==0.29.0
python-dotenv>=1.0.0
loguru>=0.7.0